Cleanup the plugin, restructure some packages.
