+++ src/main/resources/com/datapipe/jenkins/vault/VaultBuildWrapper/config.jelly
@@ -1,0 +1,9 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+    <f:section title="Vault Plugin">
+        <f:property field="configuration" header="Vault Configuration" />
+    </f:section>
+    <f:entry>
+        <f:repeatableProperty field="vaultSecrets" minimum="0" header="Vault Secret" add="Add a vault secret"></f:repeatableProperty>
+    </f:entry>
+</j:jelly>
