+++ src/main/resources/com/datapipe/jenkins/vault/model/VaultSecret/config.jelly
