--- temp/VaultBuildWrapperOld/config.jelly
@@ -1,24 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:c="/lib/credentials">
-  <f:section title="Vault Plugin">
-    <f:block>
-      <table>
-        <f:optionalBlock name="urlOverride" title="Override Vault URL?">
-          <f:entry title="Vault URL" field="vaultUrl">
-            <f:textbox />
-          </f:entry>
-        </f:optionalBlock>
-        <f:optionalBlock name="roleIdOverride" title="Override Credentials?">
-          <f:entry title="App Role Credential" field="appRoleCredentialId">
-            <c:select />
-          </f:entry>
-        </f:optionalBlock>
-      </table>
-    </f:block>
-
-    <f:entry>
-      <f:repeatableProperty field="vaultSecrets" minimum="0" header="Vault Secret" add="Add a vault secret"></f:repeatableProperty>
-    </f:entry>
-
-  </f:section>
-</j:jelly>
