--- temp/VaultBuildWrapperOld/global.jelly
@@ -1,14 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:c="/lib/credentials">
-  <f:section title="Vault Plugin">
-
-    <f:entry title="Vault URL" field="vaultUrl">
-      <f:textbox />
-    </f:entry>
-    
-    <f:entry title="App Role Credential" field="appRoleCredentialId">
-      <c:select />
-    </f:entry>
-
-  </f:section>
-</j:jelly>
