--- temp/VaultBuildWrapperOld/help-roleId.html
@@ -1,3 +1,0 @@
-<div>
-  The role ID you get from your Vault server. See https://www.vaultproject.io/docs/auth/approle.html for details. <br />
-</div>
