--- temp/VaultBuildWrapperOld/help-secretId.html
@@ -1,4 +1,0 @@
-<div>
-    The secret ID you get from your Vault server. See https://www.vaultproject.io/docs/auth/approle.html for details. <br />
-    You may want to make sure that your secret ID has no TTL and is bound to a specific CIDR / Network. Please refer to the docs for more information.
-</div>
